services/fruxAI/worker/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.0
celery>=5.3.0
